pyproject.toml
[project]
name = "stale-odds-hunter"
version = "0.1.0"
description = "Local-first prediction market trading bot for Polymarket"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "websockets>=13.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.20",
    "duckdb>=1.1",
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "streamlit>=1.40",
    "streamlit-autorefresh>=1.0",
    "pandas>=2.2",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
live = ["py-clob-client>=0.34"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
    "respx>=0.22",
]

[project.scripts]
soh = "src.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "SIM", "TCH"]
ignore = ["S101", "S104"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
